Выносит утилиты в отдельный модуль
